+++ CHANGELOG.md
@@ -8,6 +8,15 @@ All notable changes to this project will be documented in this file.
 Format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2025-02-26
+
+### 已新增 Added
+
+- 新增數值映射積木 (Arduino map)
+  Added value mapping block (Arduino map)
+- 新增數值映射相關多語系文字  
+  Added value mapping related i18n text
+
 ## [0.1.0] - 2025-02-24
 
 ### 已新增 Added